Extract markdown stuff into Markdown.js
