+++ src/components/common.js
@@ -1,5 +1,4 @@
 import _ from 'lodash/fp'
-import marked from 'marked'
 import * as qs from 'qs'
 import { Fragment, useState } from 'react'
 import { div, h, input, label, span } from 'react-hyperscript-helpers'
@@ -313,26 +312,6 @@ export const ShibbolethLink = ({ children, ...props }) => {
     children,
     icon('pop-out', { size: 12, style: { marginLeft: '0.2rem' } })
   ])
-}
-
-/**
- * WARNING: Be very careful when using custom renderers because they may override marked's built-in
- * content sanitization.
- * @param {string} children markdown content
- * @param renderers element-specific renderers
- * @param props properties for wraper div
- * @returns {object} div containing rendered markdown
- * @constructor
- */
-export const Markdown = ({ children, renderers = {}, ...props }) => {
-  const content = marked(children, {
-    renderer: Object.assign(new marked.Renderer(), renderers)
-  })
-  return div({ className: 'markdown-body', ...props, dangerouslySetInnerHTML: { __html: content } })
-}
-
-export const newWindowLinkRenderer = (href, title, text) => {
-  return `<a href="${href}" ${(title ? `title=${title}` : '')} target="_blank">${text}</a>`
 }
 
 export const IdContainer = ({ children }) => {
